crop editing

+++ resources/js/components/layout/SideMenu.vue
@@ -32,8 +32,30 @@
             <Button classes="sm icon transparent" @click="deletePlant(plant)"><Icon name="delete" colour="black"/></Button>
           </template>
         </ListItem>
+        <ListItem
+          v-show="viewing === 'crops'"
+          v-for="crop in crops"
+          :key="crop.id"
+          :highlight="crop.id === currentCrop?.id"
+          @click="selectCrop(crop)">
+          <template #image>
+            <img :src="'./images/upload/' + crop.plant.image" alt="Plant icon">
+          </template>
+          <template #title>
+            {{ crop.plant.name }} #{{ crop.id }}
+          </template>
+          <template #description>
+            <p><em>{{ crop.plant.variety }}</em></p>
+            <p>{{ monthsShort[crop.startMonth].toUpperCase() }} - {{ monthsShort[crop.endMonth].toUpperCase() }}</p>
+            <p>Qty: {{ crop.qty }}</p>
+          </template>
+          <template #actions>
+            <Button classes="sm icon transparent" @click.prevent="editCrop(crop)"><Icon name="edit" colour="black"/></Button>
+            <Button classes="sm icon transparent" @click.prevent="deleteCrop(crop)"><Icon name="delete" colour="black"/></Button>
+          </template>
+        </ListItem>
         <div class="add-button">
-          <Button class="sm icon primary" @click="createPlant"><Icon name="plus"/></Button>
+          <Button v-show="viewing === 'plants'" class="sm icon primary" @click="createPlant"><Icon name="plus"/></Button>
         </div>
       </div>
     </nav>
@@ -50,31 +72,66 @@
         </PlantForm>
       </template>
     </Modal>
+    <Modal v-if="editingCrop" @close="cancelCrop">
+      <template #header>
+        <h5>Edit Crop</h5>
+        <p>Let's edit the Crop</p>
+      </template>
+      <template #content>
+        <CropForm @done="cropSubmitted">
+          <template #buttons>
+            <Button class="primary" type="submit">Done</Button>
+          </template>
+        </CropForm>
+      </template>
+    </Modal>
+    <Modal v-if="deleteConfirm" @close="deleteConfirm = null">
+      <template #header>
+        <h5>Delete {{ deleteConfirm }}</h5>
+        <p>Are you sure you want to delete this {{ deleteConfirm }}?</p>
+      </template>
+      <template #content>
+        <p>This action cannot be undone.</p>
+      </template>
+      <template #buttons>
+        <Button class="danger outline" @click="() => {
+          if (deleteConfirm === 'plant') {
+            deletePlant(currentPlant, true)
+          } else {
+            deleteCrop(currentCrop, true)
+          }
+        }">Yes</Button>
+        <Button class="danger" @click="deleteConfirm = null">No</Button>
+      </template>
+    </Modal>
   </aside>
 </template>
 
 <script>
 import { monthsShort, monthOptionsLong } from '../../utils/consts'
-import { fetchPlants, deletePlant, createPlant } from '../../utils/api'
+import { fetchPlants, fetchCrops, deleteCrop, deletePlant, createPlant } from '../../utils/api'
 import ListItem from '../layout/ListItem.vue'
 import PlantForm from '../forms/PlantForm.vue'
-
+import CropForm from '../forms/CropForm.vue'
 export default {
   components: {
     PlantForm,
-    ListItem
+    ListItem,
+    CropForm
   },
   data () {
     return {
       monthOptionsLong: monthOptionsLong(),
       monthsShort: monthsShort(),
       viewing: 'plants',
-      mode: 'view'
+      mode: 'view',
+      deleteConfirm: null
     }
   },
   mounted () {
     this.month = new Date().getMonth() + 1
     fetchPlants(this)
+    fetchCrops(this, this.month)
   },
   computed: {
     currentPlant () {
@@ -86,6 +143,18 @@ export default {
     editingPlant () {
       return this.viewing === 'plants' && this.mode === 'edit'
     },
+    crops () {
+      return this.$store.state.crops.list
+    },
+    currentCrop () {
+      return this.$store.state.crops.current
+    },
+    editingCrop () {
+      return this.viewing === 'crops' && this.mode === 'edit'
+    },
+    cropEditingMode () {
+      return this.$store.state.crops.mode === 'edit'
+    },
     month: {
       get () {
         return this.$store.state.maps.month
@@ -95,7 +164,62 @@ export default {
       }
     }
   },
+  watch: {
+    month (val) {
+      fetchCrops(this, val)
+    },
+    viewing (val) {
+      if (val === 'crops') {
+        this.$store.commit('plants/setCurrentPlant', null)
+      } else {
+        this.$store.commit('crops/setCurrentCrop', null)
+      }
+    },
+    currentCrop (val) {
+      if (val) {
+        this.viewing = 'crops'
+      }
+    },
+    cropEditingMode (val) {
+      if (val) {
+        this.viewing = 'crops'
+        this.mode = 'edit'
+      }
+    },
+  },
   methods: {
+    selectCrop (crop) {
+      this.$store.commit('crops/setCurrentCrop', crop)
+    },
+    cancelCrop () {
+      this.mode = 'view'
+      this.$store.commit('crops/setMode', 'view')
+    },
+    editCrop (crop) {
+      this.$store.commit('crops/setCurrentCrop', crop)
+      this.mode = 'edit'
+      this.$store.commit('crops/setMode', 'edit')
+    },
+    deleteCrop (crop, confirmed = false) {
+      if (confirmed) {
+        deleteCrop(this, crop.id).then(() => {
+          fetchCrops(this, this.month)
+          this.deleteConfirm = null
+        })
+      } else {
+        this.deleteConfirm = 'crop'
+      }
+    },
+    cropSubmitted (crop) {
+      // refresh crop list
+      this.$store.commit('crops/setCurrentCrop', null)
+      this.$store.commit('crops/setCrops', [])
+      fetchCrops(this, this.month).then(() => {
+        this.$store.commit('crops/setCurrentCrop', crop)
+        this.mode = 'view'
+        this.$store.commit('crops/setMode', 'view')
+      })
+    },
     selectPlant (plant) {
       this.$store.commit('plants/setCurrentPlant', plant)
     },
@@ -117,10 +241,15 @@ export default {
       this.$store.commit('plants/setCurrentPlant', plant)
       this.mode = 'edit'
     },
-    deletePlant (plant) {
-      deletePlant(this, plant.id).then(() => {
-        fetchPlants(this)
-      })
+    deletePlant (plant, confirmed = false) {
+      if (confirmed) {
+        deletePlant(this, plant.id).then(() => {
+          fetchPlants(this)
+          this.deleteConfirm = null
+        })
+      } else {
+        this.deleteConfirm = 'plant'
+      }
     }
   }
 }
